+++ src/ContentElements/Formular.php
@@ -267,7 +267,7 @@ class Formular extends \ContentElement
 			// Ältere Anmeldung liegt bereits vor
 			// Versionierung aktivieren
 			$objVersion = new \Versions('tl_internetschach_anmeldungen', $objAnmeldung->id);
-			//$objVersions->setUsername($objMember->username);
+			$objVersion->setUsername('Internetschach-Bundle');
 			$objVersion->setUserId(0);
 			$objVersion->initialize();
 			// set-Array aktualisieren