+++ src/Resources/contao/config/config.php
@@ -27,6 +27,7 @@ $GLOBALS['BE_MOD']['content']['internetschach'] = array
 (
 	'tables'         => array('tl_internetschach', 'tl_internetschach_spieler', 'tl_internetschach_anmeldungen', 'tl_internetschach_tabellen', 'tl_internetschach_preise'),
 	'importCSV'      => array('\Schachbulle\ContaoInternetschachBundle\Classes\Import', 'importCSV'),
+	'importTable'    => array('\Schachbulle\ContaoInternetschachBundle\Classes\Import', 'importTable'),
 	'exportXLS'      => array('\Schachbulle\ContaoInternetschachBundle\Classes\Export', 'getExcel')
 );
 
@@ -38,3 +39,8 @@ $GLOBALS['BE_MOD']['content']['internetschach'] = array
  */
 $GLOBALS['TL_CTE']['includes']['internetschach_formular'] = 'Schachbulle\ContaoInternetschachBundle\ContentElements\Formular';
 $GLOBALS['TL_CTE']['includes']['internetschach_anmeldungen'] = 'Schachbulle\ContaoInternetschachBundle\ContentElements\Anmeldungen';
+
+if(TL_MODE == 'BE') 
+{
+	$GLOBALS['TL_CSS'][] = 'bundles/contaointernetschach/css/be.css'; 
+}