+++ src/Resources/contao/languages/de/modules.php
@@ -14,6 +14,10 @@ $GLOBALS['TL_LANG']['MOD']['internetschach_import_headline'] = array('Spielerdat
 $GLOBALS['TL_LANG']['MOD']['internetschach_import_file'] = array('ZIP-Datei', 'ZIP-Datei auswählen');
 $GLOBALS['TL_LANG']['MOD']['internetschach_import_submit'] = array('Import starten', 'Import starten');
 
+$GLOBALS['TL_LANG']['MOD']['internetschach_importTable_headline'] = array('Tabelle aus einer HTML- oder JSON-Datei importieren', 'Tabelle aus einer HTML- oder JSON-Datei importieren');
+$GLOBALS['TL_LANG']['MOD']['internetschach_importTable_file'] = array('Datei', 'Datei auswählen');
+$GLOBALS['TL_LANG']['MOD']['internetschach_importTable_submit'] = array('Import starten', 'Import starten');
+
 /**
  * Frontendmodule
  */