+++ src/Resources/contao/languages/de/tl_internetschach_tabellen.php
@@ -24,7 +24,10 @@ $GLOBALS['TL_LANG']['tl_internetschach_tabellen']['turnier'] = array('Turnier', 
 $GLOBALS['TL_LANG']['tl_internetschach_tabellen']['gruppe'] = array('Gruppe', 'Gruppe wählen, für die der Tabelle gilt');
 
 $GLOBALS['TL_LANG']['tl_internetschach_tabellen']['daten_legend'] = 'Ergebnisse';
-$GLOBALS['TL_LANG']['tl_internetschach_tabellen']['csv'] = array('Ergebnisse als CSV', 'Fortschrittstabelle als CSV');
+$GLOBALS['TL_LANG']['tl_internetschach_tabellen']['import'] = array('Import', 'Import');
+$GLOBALS['TL_LANG']['tl_internetschach_tabellen']['importLink'] = array('Tabelle importieren', 'Tabelle aus HTML oder JSON importieren. Vorhandene Daten werden überschrieben! Die importierten Daten werden zusätzlich im Hintergrund im Rohformat und als PHP-Array gespeichert. Mit dem PHP-Array wird bei Aktualisierungen gearbeitet.');
+$GLOBALS['TL_LANG']['tl_internetschach_tabellen']['csv'] = array('Ergebnisse als CSV (nicht editierbar)', 'Fortschrittstabelle als CSV. Die Tabelle wird entweder automatisch nach einem Import erstellt bzw. aktualisiert, wenn Sie die entsprechenden Buttons verwenden.');
+$GLOBALS['TL_LANG']['tl_internetschach_tabellen']['aktualisieren'] = array('Tabelle aktualisieren', 'Tabelle wird nach Änderung einer Option aktualisiert');
 $GLOBALS['TL_LANG']['tl_internetschach_tabellen']['disqualifikation'] = array('Disqualifizierte Spieler', 'Disqualifikationen. Mit Komma oder Bindestrich trennen.');
 $GLOBALS['TL_LANG']['tl_internetschach_tabellen']['ungewertet'] = array('Ungewertete Spieler', 'Ungewertete Spieler. Mit Komma oder Bindestrich trennen.');
 
@@ -36,3 +39,11 @@ $GLOBALS['TL_LANG']['tl_internetschach_tabellen']['intern'] = array('Interne Bem
 
 $GLOBALS['TL_LANG']['tl_internetschach_tabellen']['publish_legend'] = 'Veröffentlichung';
 $GLOBALS['TL_LANG']['tl_internetschach_tabellen']['published'] = array('Veröffentlichen', 'Tabelle veröffentlichen');
+
+// Optionen
+$GLOBALS['TL_LANG']['tl_internetschach_tabellen']['aktualisieren_optionen'] = array
+(
+	'real'      => 'Richtige Namen eintragen',
+	'prices'    => 'Gewonnene Preise eintragen',
+	'qual'      => 'Qualifikationen eintragen',
+);